fix(forms): fix permissions for commitees

+++ app/forms/models/forms.py
@@ -1,6 +1,6 @@
 import uuid
 
-from django.db import models, transaction
+from django.db import models, transaction, IntegrityError
 
 from enumchoicefield import EnumChoiceField
 from ordered_model.models import OrderedModel
@@ -16,12 +16,7 @@ from app.util.models import BaseModel
 
 
 class Form(PolymorphicModel, BasePermissionModel):
-    write_access = [
-        AdminGroup.HS,
-        AdminGroup.NOK,
-        AdminGroup.SOSIALEN,
-        AdminGroup.INDEX,
-    ]
+    write_access = AdminGroup.admin()
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     title = models.CharField(max_length=200)
     template = models.BooleanField(default=False)
@@ -47,6 +42,14 @@ class Form(PolymorphicModel, BasePermissionModel):
 
             if options:
                 field.add_options(options)
+    
+    @classmethod
+    def is_event_form(cls, request):
+        return request.data.get("resource_type") == "EventForm"
+
+    @classmethod
+    def is_group_form(cls, request):
+        return request.data.get("resource_type") == "GroupForm"
 
     @classmethod
     def has_retrieve_permission(cls, request):
@@ -67,15 +70,7 @@ class Form(PolymorphicModel, BasePermissionModel):
             return False
         if cls.is_group_form(request):
             return GroupForm.has_list_permission(request)
-        return request.user.memberships_with_events_access.exists()
-
-    @classmethod
-    def is_group_form(cls, request):
-        """
-        DRY Rest Permissions cannot handle polymorphic models
-        which requires a manual workaround for the type of form.
-        """
-        return request.data.get("resource_type") == "GroupForm"
+        return check_has_access(cls.write_access, request)
 
     @classmethod
     def has_write_permission(cls, request):
@@ -84,6 +79,8 @@ class Form(PolymorphicModel, BasePermissionModel):
 
         if cls.is_group_form(request):
             return GroupForm.has_write_permission(request)
+        if cls.is_event_form(request):
+            return EventForm.has_write_permission(request)
 
         return bool(request.user)
 
@@ -92,10 +89,8 @@ class Form(PolymorphicModel, BasePermissionModel):
         if not request.user:
             return False
 
-        if request.data.get("resource_type", "") == "EventForm":
-            event = Event.objects.get(id=request.data.get("event"))
-            return event.has_object_write_permission(request)
-
+        if cls.is_event_form(request):
+            return EventForm.has_create_permission(request)
         if cls.is_group_form(request):
             return GroupForm.has_write_permission(request)
 
@@ -119,6 +114,21 @@ class EventForm(Form):
         unique_together = ("event", "type")
         verbose_name = "Event form"
         verbose_name_plural = "Event forms"
+
+    @classmethod
+    def has_write_permission(cls, request):
+        event_id = request.data.get("event")
+        event = Event.objects.filter(id=event_id).first()
+
+        return event and event.has_object_write_permission(request)
+
+    @classmethod
+    def has_create_permission(cls, request):
+        if not request.user:
+            return False
+
+        event = Event.objects.get(id=request.data.get("event"))
+        return event.has_object_write_permission(request)
 
     def has_event_permission(self, request):
         if request.user is None:
@@ -155,7 +165,7 @@ class GroupForm(Form):
         group_slug = request.data.get("group")
         group = Group.objects.filter(slug=group_slug).first()
 
-        return request.user.is_member_of(group) or check_has_access(
+        return request.user.is_leader_of(group) or check_has_access(
             cls.write_access, request
         )
 
@@ -165,6 +175,17 @@ class GroupForm(Form):
             return False
         return check_has_access(cls.read_access, request)
 
+    def has_object_statistics_permission(self, request):
+        return self.has_object_write_permission(request)
+    
+    def has_object_read_permission(self, request):
+        # TODO: Restrict if only for group-members or draft/not public
+        return True
+  
+    def has_object_write_permission(self, request):
+        return request.user.is_leader_of(self.group) or check_has_access(
+            self.write_access, request
+        )
 
 class Field(OrderedModel):
 
@@ -201,26 +222,45 @@ class Option(OrderedModel):
 
 
 class Submission(BaseModel, BasePermissionModel):
-    read_access = [AdminGroup.HS, AdminGroup.INDEX, AdminGroup.SOSIALEN, AdminGroup.NOK]
+    read_access = AdminGroup.admin()
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     form = models.ForeignKey(Form, on_delete=models.CASCADE, related_name="submissions")
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="submissions")
-
-    class Meta:
-        unique_together = ("form", "user")
 
     def __str__(self):
         return f"{self.user.user_id}'s submission to {self.form}"
 
     @transaction.atomic
     def save(self, *args, **kwargs):
-        if isinstance(self.form, EventForm):
-            user_has_registration = not self.form.event.registrations.filter(
-                user=self.user
-            ).exists()
-            if user_has_registration:
-                Submission.objects.filter(user=self.user, form=self.form).delete()
+        print("save start")
+        existing_same_user_and_form = Submission.objects.filter(user=self.user, form=self.form)
+        print(existing_same_user_and_form.exists())
+        if existing_same_user_and_form.exists():
+            if isinstance(self.form, EventForm):
+                print("is EventForm")
+                user_has_registration = self.form.event.registrations.filter(
+                    user=self.user
+                ).exists()
+                print(user_has_registration)
+                if user_has_registration:
+                    raise IntegrityError("Du kan ikke endre innsendt spørreskjema etter påmelding")
+                else:
+                    Submission.objects.filter(user=self.user, form=self.form).delete()
+            if isinstance(self.form, GroupForm):
+                print("is GroupForm")
+                # TODO: Add can_submit_multiple to GroupForm-model
+                if not self.form.can_submit_multiple:
+                    raise IntegrityError("Dette spørreskjemaet tillater kun én innsending")
+          
         super().save(*args, **kwargs)
+
+    @classmethod
+    def _get_form_from_request(cls, request):
+        form_id = request.parser_context.get("kwargs", {}).get("form_id", None)
+        if form_id:
+            return Form.objects.get(id=form_id)
+
+        return None
 
     @classmethod
     def has_write_permission(cls, request):
@@ -238,9 +278,12 @@ class Submission(BaseModel, BasePermissionModel):
         if request.user is None:
             return False
 
-        return cls._is_own_permission(request) or check_has_access(
-            cls.read_access, request
-        )
+        form = cls._get_form_from_request(request)
+
+        if not form:
+            return False
+        
+        return cls._is_own_permission(request) or form.has_object_write_permission(request)
 
     @classmethod
     def _is_own_permission(cls, request):
@@ -254,17 +297,16 @@ class Submission(BaseModel, BasePermissionModel):
 
     @classmethod
     def has_list_permission(cls, request):
-
         if request.user is None:
             return False
 
-        return check_has_access(cls.read_access, request)
+        form = cls._get_form_from_request(request)
+        return form.has_object_write_permission(request)
 
     def has_object_read_permission(self, request):
         return self._is_own_permission(request) or check_has_access(
             self.read_access, request
         )
-
 
 class Answer(BaseModel):
